added confirmation box and add styling

+++ src/app/home/home.component.ts
@@ -4,6 +4,8 @@ import { contactBookInfo } from '../_models';
 import { ContactBookInfoService } from '../_services';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { UpdateModelComponent } from '../update-model/update-model.component';
+import { ToastrService } from 'ngx-toastr';
+import { ConfirmationModelComponent } from '../confirmation-model/confirmation-model.component';
 
 @Component({
   selector: 'app-home',
@@ -14,11 +16,13 @@ export class HomeComponent implements OnInit {
   addContactForm: FormGroup;
   contactBookData: contactBookInfo[];
   ContactBookTbl: any;
+  statusToggle: any;
 
   constructor(
     private formBuilder: FormBuilder,
     private contactBookInfoService: ContactBookInfoService,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private toastr: ToastrService,
   ) { }
 
   ngOnInit(): void {
@@ -26,28 +30,50 @@ export class HomeComponent implements OnInit {
     this.getContactBookData();
   }
 
+  /*====================Create Contact Form====================*/
   createForm() {
     this.addContactForm = this.formBuilder.group({
-      firstName: ['', Validators.required],
-      lastName: ['', Validators.required],
-      contactNumber: ['', Validators.required],
-      emailId: ['', Validators.required],
-      address: ['', Validators.required],
+      firstName: ['', {
+        validators: [Validators.required]
+      }],
+      lastName: ['', {
+        validators: [Validators.required]
+      }],
+      contactNumber: ['', {
+        validators: [Validators.required]
+      }],
+      emailId: ['', {
+        validators: [Validators.required]
+      }],
+      address: ['', {
+        validators: [Validators.required]
+      }],
+      status: [false, {
+        validators: []
+      }],
     })
   }
 
-  addContactInfo() {
-    let contactBookData = this.addContactForm.value;
-    this.addContactData(contactBookData);
-    this.addContactForm.reset();
-  }
+  /*================Add Contact Confirmation Box==================*/
+  addContactInfo(): void {
+    const dialogRef = this.dialog.open(ConfirmationModelComponent, {
+      width: '400px',
+      data: { 'msg': 'Do you want to add this contact?' }
+    });
 
-  addContactData(contactBookData: contactBookInfo) {
-    this.contactBookInfoService.addContactBookData(contactBookData).subscribe(data => {
-      this.getContactBookData();
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        let contactBookData = this.addContactForm.value;
+        this.contactBookInfoService.addContactBookData(contactBookData).subscribe(data => {
+          this.getContactBookData();
+          this.toastr.success('Contact Added Successfully!', 'Success');
+          this.addContactForm.reset();
+        });
+      }
     });
   }
 
+  /*===================Edit Contact Model Form=====================*/
   editContact(id) {
     this.openDialog(id);
   }
@@ -63,12 +89,24 @@ export class HomeComponent implements OnInit {
     });
   }
 
-  deleteContact(id) {
-    this.contactBookInfoService.deleteContactBookData(id).subscribe(data => {
-      this.getContactBookData();
+  /*===================Delete Contact Confirmation Box===================*/
+  deleteContact(id): void {
+    const dialogRef = this.dialog.open(ConfirmationModelComponent, {
+      width: '400px',
+      data: { 'msg': 'Do you want to delete this contact?' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.contactBookInfoService.deleteContactBookData(id).subscribe(data => {
+          this.getContactBookData();
+          this.toastr.success('Contact Deleted Successfully!', 'Success');
+        });
+      }
     });
   }
 
+  /*===================Get Table Data================*/
   getContactBookData() {
     this.contactBookInfoService.getContactBookData().subscribe(data => {
       this.ContactBookTbl = data;
